extract mail utilities
